create syntax_tree and split impl file
