+++ .gitignore
@@ -4,4 +4,3 @@ tmp*
 *.out
 sample*
 target/
-*.lock
