+++ mycc/src/lib.rs
@@ -1,6 +1,7 @@
 //ライブラリの宣言
 //バイナリクレートで使えるように公開する
 //::を使った階層的mod宣言はできない
-pub mod node;
+// pub mod node;
+pub mod syntax_tree;
 pub mod token;
 pub mod parser;
