+++ mycc/src/main.rs
@@ -1,8 +1,8 @@
 use std::env;
 use mycc::{
     token,
-    node,
     parser,
+    syntax_tree,
 };
 
 fn main() {
@@ -16,7 +16,7 @@ fn main() {
 
     let token = token::tokenize::tokenize(&mut argv[1]); //コマンドラインで受け取った文字列をトークン列に変換する, ここまでOK
     // println!("{:?}", token);
-    let mut code_generator = node::generative_rule::GenerativeRule::new(token);
+    let mut code_generator = syntax_tree::SyntaxTree::new(token);
     let mut node = code_generator.program();
     // println!("{:?}", node);
